@@ -1,6 +1,11 @@
 const mongoose = require('mongoose')
 
 const taskSchema = new mongoose.Schema({
+    user: {
+        type: mongoose.Schema.Types.ObjectId,
+        required: true,
+        ref: 'User'
+    },
     title: {
         type: String,
         required: true,
@@ -71,4 +76,4 @@ const taskSchema = new mongoose.Schema({
     }
 })
 
-module.exports = mongoose.model('Task', taskSchema)
+export default mongoose.models.Task || mongoose.model('Task', taskSchema)
