@@ -5,7 +5,7 @@ import { useDrop } from 'react-dnd'
 
 import { ItemTypes } from '../../lib/items-draggable'
 import { DispatchContext } from '../../pages/tasks'
-import { TASKS } from '../../state-manager/constants'
+import { TASKS_APP } from '../../state-manager/constants'
 
 const Action = ({ action }) => {
     const dispatch = useContext(DispatchContext)
@@ -14,7 +14,7 @@ const Action = ({ action }) => {
         accept: ItemTypes.TASK,
         drop: (item) => {
             dispatch({
-                type: TASKS.ADD_TASK_TO_AWARD,
+                type: TASKS_APP.ADD_TASK_TO_AWARD,
                 payload: { taskId: item.id, awardId: ''}
             })
         },
