@@ -3,11 +3,17 @@ const ADD_TASK = 'ADD_TASK'
 const SELECTED_DAYS = 'SELECTED_DAYS'
 const ADD_TASK_TO_AWARD = 'ADD_TASK_TO_AWARD'
 const ADD_DAY = 'ADD_DAY'
+const GET_BOOKS_REQUEST = 'GET_BOOKS_REQUEST'
+const GET_BOOKS_SUCCESS = 'GET_BOOKS_SUCCESS'
+const GET_BOOKS_FAILURE = 'GET_BOOKS_FAILURE'
 
-export const TASKS = {
+export const TASKS_APP = {
     MOVE_TASK,
     ADD_TASK,
     SELECTED_DAYS,
     ADD_TASK_TO_AWARD,
-    ADD_DAY
+    ADD_DAY,
+    GET_BOOKS_REQUEST,
+    GET_BOOKS_SUCCESS,
+    GET_BOOKS_FAILURE
 }
